Add common check class
